+++ crates/spacedeck/src/rabbitmq.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use url::Url;
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum Event {
+    /// Spacedeck has been initialized
+    Initialized,
+    PdfUrl(Url),
+}